Update markdown generator (submodule)
Furthermore, the markdown inserted inside `<textarea>` is now properly entity-escaped.

The previous markdown generator would output things like `&amp#124;` so that it would show up as `&#124;` in the `<textarea>` but that means if the markdown was saved to your Google Drive, it would contain the escaped entities.

There could also be other things in the markdown that would cause issues for the `<textarea>`, for example letting one of the cells contain `</textarea>`.